# java-microservice-scaffold/skeleton-service/src/test/resources/contracts/skeleton-api.openapi.yaml
openapi: 3.0.3
info:
  title: skeleton-service API
  description: 团队 Java 服务骨架契约；统一返回 Result(code, msg, data)
  version: 1.0.0

paths:
  /api/v1/health:
    get:
      tags: [health]
      operationId: health
      summary: 服务健康检查
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultHealth'

  /api/v1/users:
    post:
      tags: [users]
      operationId: createUser
      summary: 创建用户
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultLong'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      tags: [users]
      operationId: listUsers
      summary: 分页条件查询用户
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: username
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: integer
            enum: [0, 1]
        - name: email
          in: query
          schema:
            type: string
        - name: phone
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultUserPage'

  /api/v1/users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [users]
      operationId: getUserById
      summary: 按 ID 查询用户
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultUserVO'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [users]
      operationId: updateUser
      summary: 更新用户
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultUserVO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [users]
      operationId: deleteUser
      summary: 逻辑删除用户
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVoid'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  schemas:
    Result:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
        msg:
          type: string
        data:
          nullable: true

    HealthData:
      type: object
      properties:
        status:
          type: string
        application:
          type: string

    UserVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: integer
        createTime:
          type: string
          description: ISO-8601 local date-time (Jackson LocalDateTime, no timezone)
        updateTime:
          type: string
          description: ISO-8601 local date-time (Jackson LocalDateTime, no timezone)

    UserPage:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/UserVO'
        total:
          type: integer
          format: int64
        page:
          type: integer
        size:
          type: integer

    UserCreateRequest:
      type: object
      required: [username, password, email]
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        status:
          type: integer
          enum: [0, 1]
          nullable: true

    UserUpdateRequest:
      type: object
      properties:
        username:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        status:
          type: integer
          enum: [0, 1]
          nullable: true

    ResultHealth:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
        msg:
          type: string
        data:
          $ref: '#/components/schemas/HealthData'

    ResultLong:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
        msg:
          type: string
        data:
          type: integer
          format: int64
          nullable: true

    ResultUserVO:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
        msg:
          type: string
        data:
          $ref: '#/components/schemas/UserVO'

    ResultUserPage:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
        msg:
          type: string
        data:
          $ref: '#/components/schemas/UserPage'

    ResultVoid:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
        msg:
          type: string
        data:
          nullable: true

    ErrorResult:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
        msg:
          type: string
        data:
          nullable: true

  responses:
    BadRequest:
      description: 参数校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResult'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResult'
    Conflict:
      description: 数据冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResult'